% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitbit_api_oauth2.R
\name{api_auth}
\alias{api_auth}
\title{Print OAuth 2.0 URL}
\usage{
api_auth(clientID)
}
\arguments{
\item{clientID}{OAuth 2.0 client ID from Fitbit App Manager}
}
\value{
None
}
\description{
Construct string to put in GET request for authentication
}
\examples{
api_auth('23RRV4')
}

