% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitbit_api_oauth2.R
\name{get_heart}
\alias{get_heart}
\title{Fetch Fitbit heart rate data}
\usage{
get_heart(date.start, date.end, token)
}
\arguments{
\item{date.start}{First date in range to pull}

\item{date.end}{Last date in range to pull}

\item{token}{OAuth2.0 access token}
}
\value{
dataframe
}
\description{
Fetch Fitbit heart rate data
}

