% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_get.R
\name{get_nutrition}
\alias{get_nutrition}
\title{Fetch Fitbit nutrition data}
\usage{
get_nutrition(date.start, date.end)
}
\arguments{
\item{date.start}{First date in range to pull}

\item{date.end}{Last date in range to pull}
}
\value{
dataframe
}
\description{
Fetch Fitbit nutrition data
}

