% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_get.R
\name{pull_data}
\alias{pull_data}
\title{Fetch Fitbit data}
\usage{
pull_data(what, date.start, date.end)
}
\arguments{
\item{what}{Resource name to pull}

\item{date.start}{First date in range to pull}

\item{date.end}{Last date in range to pull}
}
\value{
dataframe
}
\description{
Fetch Fitbit data
}

