% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_get.R
\name{get_activity}
\alias{get_activity}
\title{Fetch Fitbit activity time series data}
\usage{
get_activity(what, date.start, date.end)
}
\arguments{
\item{what}{Resource name to pull}

\item{date.start}{First date in range to pull}

\item{date.end}{Last date in range to pull}
}
\value{
dataframe
}
\description{
Fetch Fitbit activity time series data
}

